requirements.txt
pandas
numpy
statsmodels
pyyaml
google-genai
pydantic
mcp
fastapi
uvicorn
python-multipart
